Include obb upload in apk upload
This is to ensure both uploads are in the same edit.

Also switches to "off" by default for both main and patch uploads, to
keep it working for apps without expansion files
